each index sets a different ready state and waits for all but itself to continue
now on index creating the name of the ready state is the name of the
index, and before proceding to create itself it wais for all the ready
states to be ready outside of itself.

this is done by adding a parameter to __waitForReady where one can
specify a key to exclude from the check, so if key is undefined and
excludeKey is set, waitForReady will be ready when all keys but
excludeKey are ready.

also, remove the waitForReady wrap on __getStoreProps since createIndex
already waited for readiness, on later commits all other callers to
__getStoreProps will have waitForReady set too.