move waitForReady from getStoreProps to its callers that didn't have it
this is because getStoreProps should wait for createObjectStore *and*
all it's indexes to be created before returning.

if it had waitForReady inside it when called from createIndex which set
a ready state to false just before calling getStoreProps would deadlock
since it will end up waiting for itself to create itself.

since createIndex already has waitForReady (only on createObjectStore)
waitForReady was moved to the callers of deriveKey (add and put).